requirements.txt
pytest
selenium
allure-pytest
pyyaml
pytest-html
requests
pytest-bdd